Feat: create text converter file and spec and move logic from nigh writer file
