+++ spec/text_converter_spec.rb
@@ -1,20 +1,20 @@
 require 'simplecov'
 SimpleCov.start
 require 'rspec'
-require './lib/night_writer'
+require './lib/text_converter'
 
 
-describe NightWriter do
+describe TextConverter do
 
   it 'exists' do
-    night_writer = NightWriter.new("test.txt", "braille.txt")
-    expect(night_writer).to be_a(NightWriter)
+    night_writer = TextConverter.new("test.txt", "braille.txt")
+    expect(night_writer).to be_a(TextConverter)
   end
 
   it 'constructs message to output from given arguments' do
-    night_writer = NightWriter.new("test.txt", "braille.txt")
+    night_writer = TextConverter.new("test.txt", "braille.txt")
     expect(night_writer.message).to eq("Created 'braille.txt' containing 15 characters")
-    night_writer = NightWriter.new("test.txt", "other.txt")
+    night_writer = TextConverter.new("test.txt", "other.txt")
     expect(night_writer.message).to eq("Created 'other.txt' containing 15 characters")
   end
